coffee/requirements.txt
certifi==2022.12.7
charset-normalizer==2.1.1
evdev==1.6.0
func-timeout==4.3.5
idna==3.4
paho-mqtt==1.6.1
pkcs7==0.1.2
pycryptodome==3.16.0
pynput==1.7.6
PyP100==0.1.2
python-dotenv==0.21.0
python-xlib==0.33
requests==2.28.1
six==1.16.0
urllib3==1.26.13
